extract to base class
